Refine local runtime infrastructure
