+++ tests/python/infra/instances/local/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (C) CVAT.ai Corporation
+#
+# SPDX-License-Identifier: MIT
+
+from .instance import LocalInstance, LocalPlugin
